requirements.txt
Django
mysqlclient
scrapy
scrapyd
djangorestframework
python-scrapyd-api